+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.11.0] (2023-09-11)
+
+### Added
+
+- `Player.getAudioTrack` and `Player.getSubtitleTrack` APIs to get currently selected audio and subtitle tracks
+- `SourceConfig.description` property to allow setting a description for the source
+- `Player.getThumbnail` and `Source.getThumbnail` APIs to get thumbnail images
+
+### Changed
+
+- Update Bitmovin's native Android SDK version to `3.43.0`
+
 ## [0.10.0] (2023-09-04)
 
 ### Added